Fetch ligands protocol
